Moved _rstrip_inplace to astropy.io.fits.util and added test
